SRC/requirements.txt
folium==0.14.0
numpy==1.23.5
numpy==1.24.2
pandas==2.0.1
Pillow==9.3.0
Pillow==10.0.0
plotly_express==0.4.1
streamlit==1.22.0
streamlit_folium==0.13.0
wordcloud==1.9.2
